services/user_service/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.4.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.33"}
asyncpg = "^0.29.0"
alembic = "^1.13.2"
orjson = "^3.10.7"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"
python-multipart = "^0.0.12"
sqladmin = "^0.20.0"
redis = "^5.2.0"
fastapi-limiter = "^0.1.6"
python-slugify = "^8.0.4"
flask-admin = "^2.0.0a3"
psycopg = "^3.2.3"
piccolo-admin = "^1.9.0"
fastadmin = {extras = ["fastapi", "sqlalchemy"], version = "^0.2.15"}
unicorn = "^2.1.1"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

